+++ changelog.md
@@ -6,6 +6,10 @@ This log reflects changes to the wftm documentation system. Other changes
 such as to the associated utlitilies and sample files are not tracked here.
 
 ### Log
+0.0.9
+ * Page search now supports case sensitivity
+ * Wildcards are now ? and * instead of annoying SQL _ and %
+
 0.0.8
  * Page Content field now tracks content size as you type.